requirements.txt
absl-py==2.1.0
ale-py==0.8.1
astor==0.8.1
cloudpickle==1.6.0
cycler==0.11.0
fonttools==4.38.0
future==1.0.0
gast==0.2.2
google-pasta==0.2.0
grpcio==1.62.3
gym==0.17.3
gym-notices==0.0.8
h5py==3.8.0
importlib-metadata==6.7.0
importlib-resources==5.12.0
joblib==1.3.2
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.4.5
Markdown==3.4.4
MarkupSafe==2.1.5
matplotlib==3.5.3
mpi4py==4.0.3
numpy==1.21.6
opencv-python==4.11.0.86
opt-einsum==3.3.0
packaging==24.0
pandas==1.3.5
Pillow==9.5.0
protobuf==3.20.3
pygame==2.1.0
pyglet==1.5.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
pytz==2025.1
scipy==1.7.3
six==1.17.0
stable-baselines==2.10.2
tensorboard==1.15.0
tensorflow==1.15.0
tensorflow-estimator==1.15.1
termcolor==2.3.0
typing_extensions==4.7.1
Werkzeug==2.2.3
wrapt==1.16.0
zipp==3.15.0
